add sql file tracking
