+++ .gitignore
@@ -18,4 +18,3 @@ nhs_app/resource/__pycache__/
 models/graphs/15_03_20/
 models/lite/latest_converted_model.tflite
 models/results/15_03_20/
-mysql-dev-db/ib*
