+++ mysql-dev-db/ib_buffer_pool
@@ -1,0 +1,118 @@
+10,12
+10,305
+10,120
+10,259
+10,170
+10,208
+11,16
+8,9
+8,6
+8,10
+8,15
+8,8
+105,4
+105,3
+105,2
+105,1
+85,4
+85,3
+85,2
+85,1
+83,3
+83,1
+77,3
+77,1
+74,16
+74,9
+74,8
+74,6
+74,5
+74,3
+74,1
+61,4
+61,3
+61,2
+61,1
+20,3
+20,2
+20,1
+12,3
+19,2
+19,1
+12,2
+12,1
+11,30
+11,29
+11,28
+11,27
+11,26
+11,24
+11,23
+19,3
+11,22
+11,21
+11,20
+11,19
+11,18
+11,17
+11,15
+11,14
+11,13
+11,12
+11,11
+11,10
+11,9
+11,8
+11,7
+11,6
+11,5
+11,4
+11,3
+11,2
+11,1
+10,312
+10,311
+10,301
+10,298
+10,294
+10,293
+10,291
+10,287
+10,285
+10,282
+10,275
+10,274
+10,272
+10,253
+10,250
+10,244
+10,243
+10,241
+10,240
+10,239
+10,235
+10,227
+10,216
+10,206
+10,205
+10,204
+10,201
+10,198
+10,194
+10,189
+10,187
+10,186
+10,182
+10,180
+10,177
+10,174
+18,1
+10,171
+18,3
+10,115
+10,98
+10,28
+10,1
+14,1
+10,9
+10,3
